Simplify PostProcessor::rotateFrames() interface
PostProcessor::rotateFrames() returns a (recycled) RawFrame object that can
be used to hold the next work-frame. It was required to pass a FieldType
enum value to rotateFrames() so that the returned frame could be initialized
to that value. Though that initialization can as well be done by the caller
of rotateFrames().