Remove obsolete rom/ dir in paths for ROMs
These dirs do not exist anymore for a long time. Most users will not
rely on these absolute paths but on the filepool mechanism (based on the
sha1sum of the ROM). The path mechanism itself will not be removed to
make things more comfortable for people developing ROM files for
extensions.