DirAsDSK: generalize formulas, fill-in bootsector
The DirAsDSK code handles both single and double sided disks. The (FAT12)
format for both types is slightly different: the old code got the layout right,
but it forgot to fill-in some parameters in the bootsector. This patch fixes
that.

I discovered this while investigating how to extend DirAsDSK to larger disks.
For this I had to generalize some of the calculations in the code. Now it's
relatively easy to e.g. create a DirAsDSK with 250 (instead of 80) tracks
(change one constant in DirAsDSK.cc and one in RealDrive.cc). Unfortunately it
seems the standard disk rom software can't handle such disks. E.g.
 * dskf(0) on a NMS8250 keeps reporting 713 on an (much larger) empty disk
 * dskf(0) on a turbor seems to clip at 1108
 * copying a large file (>1MB) on a turbor results in disk corruption