Replaced "make dist" implementation to use "git archive" as the base
This allows making tarballs from any tag, not just the currently
checked out tree, when called directly (not as "make dist").

Also it skips uncommitted changes and untracked files.

Upped the Python requirement to 2.6 so I could use the "except as"
syntax. I think we can actually rely on 2.7 to be present on modern
systems, but for now 2.6 is sufficient.