Replace x.begin() -> begin(x)  /  x.end() -> end(x)
In c++11 it's preferable to use the non-member begin() and end() functions over
the corresponding begin() and end() member functions. Because the former are
more general (e.g. arrays don't have member functions, but you can overload
non-member begin() and end() to work on arrays). The non-member variant is also
one character shorter ;-)

See this post for a much better explanation:
  https://stackoverflow.com/questions/7593086/why-use-non-member-begin-and-end-functions-in-c11

This patch only contains mechanical changes. (The cleanups I did while working
on this are committed as a separate patch).