Added algorithms: contains(), find_unguarded(), find_if_unguarded()
While working on the previous patch I noticed that very often we had a pattern
like the following in the code:
	auto it = find(begin(c), end(c), v);
	assert(it != end(c));
So we search for a value and then assert that the value is really found. In a
non-assert build that last check is gone, but internally in find() there still
is such a check (on each iteration). If we're really sure the value will be
found, then all those checks are redundant.

This patch implements a new algorithm find_unguarded() that only performs the
checks in assert-mode. (I did some research and 'find_unguarded' seems to be a
common name for this algorithm).

This new algorithm more clearly states the intent, and it also results in more
efficient code. E.g. the non-assert executable is now 20kB smaller. Though it
isn't used in performance critical code, so I don't expect any measurable
difference.

I also implemented a range-variant of the new find_unguarded() algorithm, in
addition to the traditional iterator-pair-variant. (C++17 will likely have
range variants for most algorithms).

--

Next to find_unguarded(), this patch also adds find_if_unguarded() and
contains(). Below is a list of transformations (simplifications) done in this
patch.

	auto it = find(begin(range), end(range), value);
	assert(it != end(range));
->	auto it = find_unguarded(range, value);

	auto it = find_if(begin(range), end(range), predicate);
	assert(it != end(range));
->	auto it = find_if_unguarded(range, predicate);

	find(begin(range), end(range), value) != end(range)
->	contains(range, value)

	count(begin(range), end(range), value) == 0
->	!contains(range, value)

	find_if(begin(range), end(range), predicate) != end(range)
->	any_of(begin(range), end(range), predicate)

	find_if(begin(range), end(range), predicate) == end(range)
->	none_of(begin(range), end(range), predicate)