Made display_deform=3d effect openGL ES friendly
Compared to plain openGL, openGL ES 2.0 removed quite some (legacy)
functionality. The goal of this and the next patches is to rewrite our current
code in the common subset of openGL and openGL ES 2.0. So initially the code
remains openGL, but it should make the switch to openGL ES later easier.

This patch focusses on the 'display_deform=3d'-effect. I started with this one
because (from an openGL mathematics pov) it's the most interesting.

There are 4 main changes in this patch:
- Replaced openGL matrix manipulation calls
    glMatrixMode(), glPushMatrix(), glPopMatrix(), glLoadIdentity(),
    glFrustum(), glTranslate(), glRotate(), glScale(), ...
  with 'manual' matrix calculations (see previous patch).
- OpenGL ES no longer support display lists. These should be replaced with
  glDrawArrays()/glDrawElements (+ optionally vertex buffer objects (VBO), see
  next patch).
- Direct drawing mode (using glBegin()/glEnd()) is no longer supported. Should
  also be replaced by glDrawArrays()/glDrawElements().
- The fixed-function rendering pipeline is removed in openGL ES 2.0, this means
  that for every drawing operation you need to write a (small) vertex and
  fragment shader. In particular for this effect it means we must 'manually'
  program the lighting calculation.

This effect was the most intensive user of matrix manipulations, so I'm
confident that the remaining code that still uses these functions can also
easily be replaced (point 1). This effect was the only (remaining) user of
openGL display lists, so point 2 is done. Points 3 and 4 still need to be fixed
in a lot of other locations.