Use vertex buffers objects (VBO) in display_deform=3d effect
In the display_deform=3d effect we approximate the curved monitor by a mesh of
16x16x2 triangles. Times three for the number of vertices, and each vertex
consists out of a 3D position, a normal vector and a texture coordinate. All
this data remains unchanged from frame-to-frame, but (before this patch) each
frame we do transfer all this data from the CPU to the video card.

To make this transfer more efficient, openGL(ES) introduced Vertex Buffer
Objects (VBO). This way the data can be uploaded to the video card once and be
reused whenever it's needed (each frame in this case). The openGL ES coding
guidelines suggest to use VBOs whenever possible. Desktop machines have no
problem drawing this effect using the old way, but it might be important on
embedded devices (even if only to reduce power consumption).