Made noise effect openGL ES friendly
Another post-processing effect that is applied in GLPostProcessor is the
'noise' effect. This patch changes the code so that it only uses openGL
functions that are also available in openGL ES 2.0.

Like in a previous patch we now have a write our own shaders to replace the old
fixed-function rendering pipeline. Before vertex coordinates were transformed
by the global model-view-projection matrix, in openMSX this matrix was setup so
that xy-vertex coordinates directly correspond to host pixels. Now we have to
do this transformation ourselves. I choose not to transform the incoming vertex
coordinates at all, IOW they already must be clip-space-coordinates (stuff
inside the axis-aligned cube with sides at {+1,-1}{x,y,z} is shown, the rest is
clipped). This has the advantage that the effect is now independent of the host
screen resolution (before the coordinates had to be multiplied by the scale
factor).

This patch adds an extremely simple shader program. It's likely that future
patches will need a very similar or even identical program. So this program may
be shared and/or (slightly) generalized later.