Made glow effect openGL ES friendly
Similar to the previous patch. This changes the fixed-function implementation
of the glow effect into a version that only uses openGL functions that are
also available in openGL ES. This required writing a custom shader program.

This patch also changes one other drawing call in the GLPostProcessor code (the
case when display_deform=normal and glow!=0). So now all glbegin()/glEnd()
functions are removed from the GLPostProcessor code.