Made GLSnow code openGL ES friendly
The snow effect (set power off) is similar to the noise effect, so it also
isn't surprising that the same shader program could be used. This patch reuses
the source code of that shader but still creates a 2nd instance of that
program. This will be changed in a later patch.