Unify and generalize 'glow' and 'tex2D' shader programs
The two recently created shader programs 'glow' (the glow effect) and 'tex2D'
(draw a simple texture, already used in a few places) were _very_ similar
programs. The only difference was that 'glow' still multiplied the texture
values by an alpha-factor. But if we set alpha=1.0 they are identical.

The next patch in this series will again need a _very_ similar program, the
only difference is that the vertex positions need to be transformed. But if
we pass the identity matrix as transformation-matrix, the programs again
become identical.

There are multiple views on how many unique/shared shader programs to make.
The two extreme views are:
 a) For every drawing operation create an as specific program as possible
 b) Create an as general program as possible
Point a) has the advantage that the program will likely execute faster, but
it has the disadvantage that you'll often need to switch between programs
(because such a switch is relatively expensive). The general advice (given in
various openGL fora) seems to be "be pragmatic", so don't go for a) nor b) but
somewhere inbetween.

This patch adds a new file 'GLPrograms'. This will hold a collection of
generally reusable programs, though ATM it only contains one such program.