Move GLUtil stuff from 'openmsx' to 'gl' namespace
We currently only use one namespace in all the openMSX source code (*). Namely
the 'openmsx' namespace. Often also generic utility functions that would
perfectly be reusable in other projects are put inside this 'openmsx'
namespace. But if we're only using only one namespace we can just as well use
no namespaces at all.

This patch moves the functions in the GLUtili.{hh,cc} files from the 'openmsx'
to the 'gl' namespace. It also adjusts the namespace qualifiers in the other
code. So all these are only mechanical changes (no changes in functionality).

(*) Actually the first paragraph is not strictly true. There are already
    some places were we do use more meaningful namespaces. This patch is
    another (small) step in that direction.