Allow null openGL object handles
Our class gl::Texture is a wrapper around an openGL texture handle. This handle
itself is just a 'GLuint' type (32-bit integer). Our Texture class is
responsible for properly allocating and deallocating the handle. It also offers
some convenience functions (e.g. switch between nearest-neighbour or linear-
interpolation texture sampling).

The value '0' for the openGL handle is special, it represents a 'not-allocated'
texture. Before this patch our Texture class could not represent this value (it
would always allocate a texture on construction) but this 0-value is sometimes
very useful.

Actually we already (wanted to) use this 0-value in our code:
- SDLGLOutputSurface sometimes does or does not need a texture. As a workaround
  we wrapped the Texture object in a unique_ptr which can be nullptr. This of
  course works, but think about what we're doing: we have a pointer to a heap
  allocated object that contains a single integer (almost seems like Java ;-)
- The GLImage class also only sometimes needs a Texture. Instead of wrapping in
  unique_ptr, here we simply did not use our Texture class, instead we directly
  used the low level openGL functions to manage the texture ourselves.

So this patch allows to create null Texture handles and it adjusts the
SDLGLOutputSurface and GLImage code to make use of this. Similarly this patch
also allows to create null ShaderProgram objects, but that functionality is
not yet used (see next patch).

--

Another critique to our openGL handler classes can be read here:
   https://www.opengl.org/wiki/Common_Mistakes#The_Object_Oriented_Language_Problem
Very short summary: sometimes wrapper classes hide the underlying openGL
function calls too much, and this makes it too easy to write inefficient code
(e.g. bind a texture when it was already bound). Or wrong code (unintentionally
modify global state).

I believe this critique has some truth in it, but this patch does not try to
address it (yet?).