Remove limitation of #blocks/sectors for flash
In addition, process the reset command as any other commands. Without
this, the Quadruple Byte Program command will fail when it is
programming a byte 0xF0.

Previous implementation only allowed 32 bits for the write protect
flags, so in fact it only supported up to 32 blocks/sectors. Removing
this limitation allows to use AmdFlash also for MegaFlashROM SCC+ SD,
which uses 135 blocks.