Implement element-wise vector division
Some of my work-in-progress patches used this, though in the final version that
operation was no longer needed. I'm committing it anyway: might be useful again
in the future (and it has proper tests). Because it are all inline functions,
having these unused operations in the code doesn't make any difference in the
final executable.