Use glDrawArrays() in scalers algorithms
The last openGL(ES) patch series eliminated all but one call(s) to
glBegin()/glEnd(). This patch removes the last one. Remember openGL ES doesn't
support immediate drawing mode anymore.

So this patch uses glDrawArrays(). For this we need to pass the vertex-position
and -texture coordinates in custom attributes instead of pre-defined openGL
attributes(*). So this patch also changes all the vertex programs to use
  a_position  instead of  gl_Vertex
  a_texCoord  instead of  gl_MultitexCoord{0,1}
Note that we use only one (3D) attribute to pass two sets of (2D) texture
coordinates. This is possible because in our case both coordinates always share
the same x-component.
(*) It might be possible in plain-openGL to pass stuff for built-in-attributes
    via glDrawArrays() (although I don't know how), but for sure it's not
    possible for openGL ES. Also built-in attributes don't have any advantage
    over custom attributes.

We also change the built-in uniforms to custom uniforms. Again because openGL ES
doesn't provide them anymore. So
  u_mvpMatrix  instead of  gl_ModelViewProjectionMatrix
(though most often this was hidden inside a ftransform() call).

This patch introduces some repetition in the setup code for the different
scalers. That will be eliminated again later in this patch series.