Make GLDefaultScaler more similar to the other scalers
The previous patch introduced the new GLDefaultScaler class, but it kept the
corresponding shader program(s) unchanged. It has one shader program to handle
the superimpose case and it reuses the generic texture-shader-program for the
non-superimpose case. All other GLScaler subclasses use only one program for
both case (actually one source code that gets compiled twice with a different
define).

This patch renames the shader program from 'superImpose' to 'default' (because
it now also handles the non-superimpose case). But more important it makes the
C++ code for GLDefaultScaler more similar to the other classes. And that's
useful for the next patches.