Share code in GLScaler subclasses, scaleImage() part
This patch continues the common code sharing of the previous patch. Now the
(C++ part of) the actual shading code is handled.

To make the code even more similar (and make more sharing possible) we also
tweaked the vertex shader code a bit: many vertex shaders have a 'texSize'
uniform. Often this was a 2D vector (X and Y dimension of the 'MSX' texture),
but sometimes it was a 3D vector (X, Y as before but an additional Y for the
superimpose texture). Now we always pass a 3D vector (it doesn't hurt to have
unused components). This 'texSize' uniform (the first 2 components) was always
initialized in the same way, except in the HQlite scaler. Though with some
refactoring this difference could be eliminated (and hq and hqlite become even
more similar).