Fixed invalidateMemCache() bug in MSXSCCPlusCart
There was a missing call to invalidateMemCache() in case the SCC+ cartridge was
switched between SCC+/SCC-compatibility/SCC-disabled mode. This bug has been
present since 2c273d8 (2012) but it never(?) triggered. I guess because
switching this mode is not a very frequent operation.

There was also a invalidateMemCache() performance bug (not correctness bug)
when switching the enabled RAM segments, the memory was kept uncacheable for
longer than strictly needed.

Anyway, both are fixed now while at the same time simplifying the code.

The KonamiSCC code does not have this problem. TODO verify the other devices
that contain a SCC.