Fix 4635b13 more cleanly
We recently introduced the file GLPrograms.{hh,cc}. This was initially meant as
a collection of openGL shader programs that can replace the old fixed-function
openGL programs. Though we quickly added non-shader-program stuff like a global
ModelViewProjection matrix and the shared fallback scaler. So this patch
renames GLPrograms to GLContext.

The underlying problem addressed in commit 4635b13 is actually one of object
lifetimes. When the openGL context is destroyed we used to call the (old)
destroyprograms() function and that one would logically destroy GLPrograms. But
because it's not implemented via a c++ destructor, later c++ 'really' wants to
destroy the object and we get into the situation described in 4635b13. This
patch models gl::Context with an actual object (with matching lifetime) and
thus the problem disappears.

(detail) One disadvantage of this new approach is that on each access to
gl::context we now have an extra indirection. Can be avoided by using
placement-new and manual destruction, but let's not do that. The extra cost is
small enough and it's not performance critical anyway.