Added circular_buffer::push_back(initializer_list)
This patch adds two new methods:
  circular_buffer<T>::push_back(initializer_list<T>)
  cb_queue<T>::push_back(initializer_list<T>)
This allows to insert multiple values at once with a syntax like:
  myQueue.push_back({1,2,3,4,5});
This is (or will be) useful in the (in development) SdCard class.

This requires support for c++11 initializer-lists. Gcc and clang support
those. According to the documentation visual studio should also support
them, but let's see.