Fixed newline stuff in Base64
Commit 178edb626ed accidentally broke the insertion of newline characters in
base64 encoded strings. This wasn't really a problem because when decoding
base64 strings newlines are completely ignored. Though the code didn't behave
as it was intended.

Found while writing a unittest for the base64 class. For now those unittest
still live on my personal (local) branch. The build system still needs to be
changed before those tests are more generally useable.

This patch also changes the output format (line length) to be identical to
the output of the coreutils 'base64' tool.