Fix trailing space cosmetic issue in HexDump
Found while writing a unit test for this class. When the input is not an
integer multiple of 16 bytes, the output has a trailing space. Not real problem
because whitespace is ignored when decoding such strings. Also the current code
doesn't use the encoding part anymore (it can still be enabled by making small
changes in the source code and recompiling, sometimes useful to debug stuff).