Added initial SD card support for MegaFlashROM SCC+ SD
This implementation is fully working for the happy flow. A list of TODO
items is in SdCard.cc. Next patches will address these issues. As this
version is already working well, I'm committing it to more people can
help testing.

Thanks a lot to Manuel Pazos and Wouter for their support!