Revert da70820d [6/13] map->vector: Tiger tree cache
The elements in a vector are relocated when the vector grows, so it's not safe
to return references to them. See comments in patch for more details.

This bug could trigger when using a IDE or SCSI interface with more than one
attached harddisk (our default extensions all only have one HD).

(I also reviewed the 12 other patches of the patch serie containing da70820d,
those all seem fine.)