Cleanup sha1 class
The sha1 class had a method calcWithProgress() that reports progress
information while calculating the sha1 of large files. Though to do this it
depends on some very openmsx-specific classes (CliComm and EventDistributor),
so it doesn't really belong in the openmsx-agnostic sha1 class. There was only
one user of this functionality (FilePool), so I moved it there now. If there
are more users in the future we can still factor it out.

Sha1Sum offered operator== and operator<, for consistency I also implemented
!=, <=, > and >=. Even though the current code doesn't need them. I also added
operator<<(ostream&, Sha1Sum&). This is useful to report errors in a unittest.
All these new functions are inline, so they don't cost anything when not used.

All these changes were made while working on a unittest for sha1.cc. This test
itself is not yet committed. (Hopefully soon.)