Generalize KeyRange utility
KeyRange could be used like this:
	map<T1, T2> m;
	for (auto& e : keys(m)) {...
So it could transform a std::map into a range of all the keys in that map.

This patch generalizes the keys() function so that it also works on e.g.
a vector<pair<T1, T2>> or even vector<tuple<T1, T2, ...>>. It also adds a
values() function (that gives a range of the values in a map, or the 2nd
element in a pair/tuple). And it adds a elements<N>() function to pick the
N-the element out of a collection of tuples.

This patch also makes this utility more STL-compliant (e.g. the iterators can
now also be used in algorithms like 'distance').

I made both these changes while working on a unittest for this utility.