Fixed bug#541: watch points affect execution flow
There was a bug in MSXCPUInterface::writeMem(), writes to regions with the
'disallowWriteCache' flag set would get executed twice (more or less, see code
for details).

The disallowWriteCache flag is set for the following regions:
 - The secondary subslot register (address 0xffff)
 - Writes that ignore the slot-select signal (only used by Lode Runner IIRC)
 - Watchpoints

So for bug#541 when a (write-)watchpoint was set on the flashrom, the flash
received the write twice, and this ruins the specific write-sequence needed to
trigger flash commands.

This bug was also visible in the secondary slot select register. E.g. on a
machine with a memory mapper in an expanded slot (e.g. a turboR) execute this
in MSX-BASIC:
  OUT &hFE,0      ' select same mapper segment in page 2 and 3
  ? PEEK(&hBFFF)  ' gives some value
  POKE &hFFFF,0   ' write to secondary slot select register
  ? PEEK(&hBFFF)  ' should give the same value as above, but before this
                  ' patch in openMSX it gave the value 0