Remove gl::LuminanceTexture
There was only one user of this class left: the GL noise effect. (The old SDLGL
renderer used this class as well). But we can just as well use the generic
Texture class. This patch by itself is a small cleanup, but it's mostly meant
as a preparation for the next patch.

This patch makes 2 other small changes:
- It enables linear interpolation for the noise effect. The difference is very
  subtle but IMHO it looks better.
- As internal format use GL_LUMINANCE instead of GL_LUMINANCE8. OpenGL ES
  doesn't support the latter, and it doesn't really matter anyway.