Make texture use GL_CLAMP_TO_EDGE by default
The default openGL texture wrapping mode is GL_WRAP. And our Texture class kept
this default. Though in openGL ES wrapping is only supported for power-of-two
sized textures (makes sense from a hardware point of view: AND versus MODULO).
But many of our textures have arbitrary dimensions. Luckily npot-textures are
supported in openGL ES with the GL_CLIP_TO_EDGE wrapping mode.

For the majority of our textures we don't need (or even want) wrapping. But we
didn't bother specifying this because it doesn't matter (we never access the
texture with texture coordinates outside [0,1]). This patch changes the Texture
class so that by default it sets the GL_CLIP_TO_EDGE wrapping mode.

There are only two locations in the code that require wrapping textures: the
noise and the snow effect (I hope I didn't miss any). And both use
pot-textures.