Implemented optimized StringOp::trimLeft()
Optimized for the case we're only trimming a single character as opposed to a
whole range of characters (e.g. only trim spaces instead of all whitespace
characters).

Found while working on a unit test for StringOp. I mainly added this to make it
symmetric with the trimRight() functions which already have this optimization.