Allow to rename machine-specific commands
See this forum topic:
  http://www.msx.org/forum/msx-talk/openmsx/openmsx-misc-tcl-questions-topic

The problem occurred with MSX machine specific commands (like
'cassetteplayer'). OpenMSX allows to have multiple MSX machines in memory. Not
all those machines have the same set of commands (e.g. a turbor doesn't have a
'cassetteplayer' command, or some machines may (not) have a 'diskb' command).
All those machine specific commands reside in a per-machine namespace. For
example machine1::cassetteplayer or machine2::diskb (this allows to execute the
command in the correct machine in case there are multiple machines in memory).
Though in the vast majority of the cases the user wants to execute commands in
the 'current' machine, so for all machine specific commands there's also a
'proxy' command with the same name in the global namespace. And if you invoke
the global 'cassetteplayer' command, it will delegate to
[machine]::cassetteplayer. (The 'machine' command without parameters returns
the ID for the currently active machine).

Now what happens if you execute
  rename cassetteplayer tapedeck
This will (only) rename the global proxy command (not also all the machine
specific commands). Before this patch there was actually only one such proxy
command that did the delegation for all commands, using something like
   [machine]::$arg0
So it would e.g. try to execute machine1::tapedeck. And that's indeed an error.

This patch replaces the single proxy command with one per unique command name.
And those new proxy commands remember the original name of the command. So the
proxy for 'cassetteplayer' always delegates to 'machine<x>::cassetteplayer',
irregardless of whether the proxy itself was renamed or not.

--

Note that the rename trick will still go wrong with a smarter catapult. With
smarter I mean one that knows about the multiple machines and directly invokes
commands in the machine specific namespaces.