[1/7] remove TclObject::getInterpreter()
The goal of this patch series is to speed-up Tcl-script heavy scenarios. Part
of that is to make our TclObject class more light-weight. We do that by
(eventually) removing the Tcl_Interp* member variable from TclObject. Though
this patch, as a smaller preparation step, only removes the getInterpreter()
method from TclObject.

Next to removing getInterpreter() this patch also adds an Interpreter&
parameter to the TclObject::executeCommand() method. (Might have been possible
to split the change of this and the previous paragraph in two patches, though
these often occur together in the code, so I didn't bother).

Because it's no longer possible to query the Interpreter via a TclObject, we
now sometimes have to store/pass an Interpreter object next to a TclObject.
So temporally this is more heavyweight, only later in this series it becomes
a netto-win.