[2/7] pass Interpreter& param to TclObject methods
This continues the work of the previous patch: it (only) prepares the way to
remove the Tcl_Interp* member variable from the TclObject class. We now
explicitly pass an Interpreter& parameter to all the TclObject methods that
need it.

The changed methods are:
  getInt(), getBoolean(), getDouble(), getListLength(), getListIndex(),
  getDictValue(), evalBool() and (already in previous patch) executeCommand()
With the exception of the latter two (evalBool() and executeCommand()) the
Interpreter parameter is only required to be able to produce a decent error
message in case something goes wrong (e.g. the current value of the TclObject
cannot be interpreted as an integer). That explains why the two methods
getString() and getBinary() didn't receive this extra Interpreter parameter
(any value can be interpreted as string or as binary).

Before this patch, the Interpreter could _optionally_ be set via the TclObject
constructor, now it's a required parameter for the getter methods. I'm not
aware of any code where we actually forgot to pass the Interpreter via the
constructor but the interface allowed this mistake. So with this patch I'm
more confident that all getters are now able to produce good error messages.