[4/7] Replace vector<TclObject> with array_ref<TclObject>
This is the most important change of this patch series. It eliminates the
overhead we introduced by wrapping the Tcl C-API into our C++ code. The main
change is located in Interpreter::commandProc(). Almost all other changes made
in this patch are mechanical changes to replace vector<TclObject> with
array_ref<TclObject>.

When the Tcl library executes a command implemented in C, it calls a function
with (among others) these two parameters: int objc, TclObj** objv (thus similar
to the two classic 'main(int argc, char** argv)' parameters). Before this patch
series we would then do the following:
 - Reserve space in a vector<TclObject>.
 - Wrap each of the TclObj's into a TclObject (wrapper = TclObj* + TclInterp*).
 - This wrapping includes incrementing the reference count of the TclObj.
 - Call our Command::execute(const vector<TclObject>&) method.
 - Destroy all TclObject's, this means decreasing the reference counts again.
 - Deallocate the vector.

Now that TclObject is a very thin wrapper around TclObj* (actually so thin that
TclObject and TclObj* have an identical binary representation) we can eliminate
most of the steps in the above paragraph. We now do:
 - Reinterpret the array of TclObj*'s as an array of TclObject's (allowed
   because both have the same binary representation). And wrap objc+objv in an
   array_ref. Internally array_ref contains a pointer + size, so this step is
   basically a nop.
 - Call our Command::execute(array_ref<TclObject>) method.
So compared to the old situation we've eliminated one heap allocation/
deallocation and one increment-ref/decrement-ref/ptr-copy for each argument.

A 'const vector<TclObject>&' and an 'array_ref<TclObject>' parameter have an
almost identical interface. So except for mechanically changing the type of the
parameter this change requires no changes in the implementation of all our
commands.

Next to the above array_ref has 2 other minor advantages over vector:
 - Allows to pass a stack- (instead of heap-) allocated collection (e.g. see
   DiskCommand::execute()).
 - Allows to directly pass a subrange of an existing collection instead of
   having to make a copy (e.g. see AviRecorder::processToggle()).