[6/7] Remove vector<string> interface from RecordedCommand
The Command class (and related classes) offers 2 interfaces for the same thing:
  void execute(array_ref<TclObject> tokens, TclObject& result);
  string execute(const vector<string>& tokens);
A subclass needs to implement exactly one of these two interfaces. The former
will result in faster code, the latter might be more convenient to use.

Though this dual interface also has some disadvantages:
- C++ doesn't have a way to enforce 'reimplement exactly one of these two', so
  mistakes are only caught at run-time instead of compile- (or link-)time.
- Most command implementations are initially copy-pasted from another already
  existing command. Sometimes picking the wrong (less appropriate) interface.

Also the initial motivation (easier to use) for this dual interface has
diminished over time. It's mostly a matter of adding 'getString()'. And for the
cases where you actually need an int/bool/double the TclObject interface is
already easier to use.

So this and the next patch will remove this dual interface and only keep the
fast (TclObject) variant. This patch changes the 'RecordedCommand' class.