[7/7] Remove vector<string> interface from Command
Similar to previous patch but now in Command class (instead of
RecordedCommand class).

This patch also changes the type of some parameters in various functions from
const string& to string_ref. Otherwise we still pay for a string construction
and (part of) the reason for this patch was to avoid such conversions.