Tweak string passing in src/console/
Found this when cleaning up some local git branches I had (hence the old commit
date).

Prefer to pass string-like parameter by std::string-value instead of 'const
std::string&' or 'string_view' for sink-parameters. That is: parameters that
are unconditionally copied in the implementation of the function. Before this
change this always required making a string-copy (in the callee). After this
change the callee can move the argument and the caller either a) makes a copy
or b) performs a move. In case of a) the amount of copies remains the same, in
case of b) we avoid the copy.