Revert "[5/7] use std::initializer_list in TclObject"
This reverts commit 312b7551ed5107797b5bccbc20a586a517d04ec4.

See bug#542 Fails to start up from 183-g312b755-windows-vc-x86 (on Windows Visual Studio 32-bit build)
http://sourceforge.net/p/openmsx/bugs/542/

I reviewed the code (the change related to the FilePool setting) and I'm fairly
sure it's correct. So a bug in the (32 bit) visual studio compiler??

For now I'll revert this change. We should try again when a new visual studio
version is released.