[4/6] Added SSE2 optimization in draw8()
Use SSE2 instructions to draw 4 pixels at-a-time (32bpp). It would be possible
to also draw 8 pixel at-a-time for 16bpp. Though on x86 16bpp is likely not
often used anymore, so this is not worth it.

I performed the following benchmark:
- run first 100 seconds (emutime) of 'Maze of Galious'
- scale_factor=1, horizontal_stretch=320, maxframeskip=0,
  sound_driver=null, throttle=off
I used callgrind (with cache/branch simulation enabled) to get an estimate for
the cycle count of the total program and for the renderGraphic2() method. In
the above scenario the renderGraphic2() is the 2nd most expensive function!!!
Even more expensive than CPU emulation. (Only Scaler1 is more expensive).

All patches in this series before this one had no impact on speed. This patch
itself reduce the total cycle count to 96% of the original, for renderGraphic2()
the cycles are reduced to 88% of the original.