[5/6] Specialized version of renderGraphic2()
Revision 8651f1521 made renderGraphic2() a tiny bit slower because there can
now also be mirroring in the pattern table. This situation can only happen on
TMS9928 and only for very specific MSX software, so only very rarely in a
'typical' use case.

Additionally the renderGraphic2() routine already had to deal with
- Mirroring in the color table
- V9958 horizontal scroll
Also these two features rarely trigger for typical users.

In isolation these 3 features only have a small cost, but combined we get to
the following numbers (one loop iteration draws 8 pixels):
- Full-featured inner loop: 43 instructions (196 bytes)
- Stripped down inner loop: 30 instructions (140 bytes)
(numbers for gcc-4.8, x86_64, 32bpp, SSE2).

So this patch first checks whether it's possible to use the fast loop (it often
is) and then uses either the faster or slower variant.

I performed the same benchmark as described in the previous patch, compared to
before this patch series, the total cycle count is now reduced to 94% of the
original, the cycle count for renderGraphic2() is reduced to 79%. (So roughly a
similar incremental speedup as in the previous patch!)