[6/6] Alternative (faster) isContinuous() method
Before this patch series isContinuous() was only used inside getReadArea() in
an assert (so not used at all in an optimized build). It is a generic routine
that can handle all corner cases.

The previous patch in this series started using isContinuous() also in
non-assert code. It always tests blocks that are 0x800 bytes in size and
are also aligned at 0x800 bytes. For such special cases it's possible to
implement a much faster version of isContinuous().

Benchmark (same as described before) showed only a small additional
improvement. Numbers for this whole patch series:
- total emulation   performance: 94% of original (1.06x faster)
- rennderGraphic2() performance: 78% of original (1.28x faster)