Cache the value of the cmdTiming setting
While working on the previous patch series I noticed that string comparisons
were used suspiciously many times. They even took almost 2% of total emulation
time in that particular benchmark! After more investigation this was because on
every CPU->VRAM write we check the value of an enum setting, and this involves
some string comparisons.

Fixed by caching the value in a boolean member variable and listening for
changes on this setting (the setting rarely changes).

I probably didn't notice this inefficiency before for one of these reasons (or
a combination of these):
- Before the Setting class simplification querying enum settings was more
  efficient.
- Before the VRAM access slots refactoring we didn't need to check a setting
  on every VRAM read/write.
- I mostly benchmarked MSX2 games (this benchmark was MSX1), maybe those games
  (mostly) used the VDP command engine to modify the VRAM content, so CPU->VRAM
  writes weren't used that much.