[3/2] Optimize left/right border drawing
Ok, one more patch in this series than I originally anticipated ;-)

In the RawFrame class we store all (horizontal)border rendering settings and
also one bit per line to indicate whether it contains horizontal border pixels
or not. Though the latter info is actually redundant: we already store a
'line-width' for each line in the frame. Line width == 1 means it's a vertical
border line, 320/640 (or still other value for v9990) means a MSX high or low
resolution line.

It turns out that line-width == 1 or != 1 is the same information as the bit we
would set for 'contains horizontal border'. So we don't need to store those
bits.

This patch removes those bits. It also changes some details in when exactly we
update line-width, so that the value of line-width of the previous frame
remains valid for as long as we need it.

Note that (as a much older optimization) for vertical border lines we just
store a single pixel, not a whole line containing the same repeated color. So
contrary to what you might expect at first sight, it's not possible to reuse
the pixels of a vertical border line for a horizontal border.